[1418] Cached report refactor - Moved broken link reports here from ckanext-dgu. Register reports using new 1418 infrastructure.

+++ ckanext/archiver/model.py
@@ -96,7 +96,7 @@ class Archival(Base):
     # Details of last successful archival
     cache_filepath = Column(types.UnicodeText)
     cache_url = Column(types.UnicodeText)
-    size = Column(types.Integer, default=0)
+    size = Column(types.BigInteger, default=0)
     mimetype = Column(types.UnicodeText)
     hash = Column(types.UnicodeText)
